--- hmr.ts
@@ -1,61 +1,0 @@
-// Copyright 2022 the Deno authors. All rights reserved. MIT license.
-
-const HMR_CLIENT = `let socket;
-let reconnectTimer;
-
-const wsOrigin = window.location.origin
-  .replace("http", "ws")
-  .replace("https", "wss");
-const hmrUrl = wsOrigin + "/hmr";
-
-hmrSocket();
-
-function hmrSocket(callback) {
-  if (socket) {
-    socket.close();
-  }
-
-  socket = new WebSocket(hmrUrl);
-  socket.addEventListener("open", callback);
-  socket.addEventListener("message", (event) => {
-    if (event.data === "refresh") {
-      console.log("refreshings");
-      window.location.reload();
-    }
-  });
-
-  socket.addEventListener("close", () => {
-    console.log("reconnecting...");
-    clearTimeout(reconnectTimer);
-    reconnectTimer = setTimeout(() => {
-      hmrSocket(() => {
-        window.location.reload();
-      });
-    }, 1000);
-  });
-}
-`;
-
-export const HMR_SOCKETS: Set<WebSocket> = new Set();
-
-export function hmrHandler(req: Request) {
-  const { pathname } = new URL(req.url);
-
-  if (pathname == "/hmr.js") {
-    return new Response(HMR_CLIENT, {
-      headers: {
-        "content-type": "application/javascript",
-      },
-    });
-  }
-
-  if (pathname == "/hmr") {
-    const { response, socket } = Deno.upgradeWebSocket(req);
-    HMR_SOCKETS.add(socket);
-    socket.onclose = () => {
-      HMR_SOCKETS.delete(socket);
-    };
-
-    return response;
-  }
-}
